# openapi.yaml
openapi: 3.1.0
info:
  title: OssPilot Tenant API
  version: 0.1.0
paths:
...
  /api/buckets/{bucket_name}/objects:
    get:
      summary: List object metadata
      responses:
        "200":
          description: items and prefixes
    delete:
      summary: Soft-delete objects into .trash/, or permanently delete when permanent=true
      responses:
        "200":
          description: deleted and failed keys
        "503":
          description: storage is not configured
...
  /api/buckets/{bucket_name}/trash:
    get:
      summary: List soft-deleted objects (original keys, without .trash/)
      responses:
        "200":
          description: items
    delete:
      summary: Permanently delete objects from the recycle bin
      responses:
        "200":
          description: succeeded and failed keys
        "503":
          description: storage is not configured
  /api/buckets/{bucket_name}/trash/restore:
    post:
      summary: Restore objects from .trash/ back to their original keys
      responses:
        "200":
          description: succeeded and failed keys
        "503":
          description: storage is not configured
...
  /api/share-links:
    get:
      summary: List active share links
      responses:
        "200":
          description: items
    post:
      summary: Create a temporary share link
      responses:
        "200":
          description: item
        "503":
          description: storage is not configured
  /api/share-links/{link_id}:
    delete:
      summary: Revoke a share link
      responses:
        "204":
          description: revoked
        "404":
          description: not found
  /s/{token}:
    get:
      summary: Public share access (presigned URLs, no JWT)
      responses:
        "200":
          description: download and preview urls
        "401":
          description: invalid or missing password
        "403":
          description: access limit reached
        "404":
          description: not found
        "410":
          description: expired
...